refactor(http): split http_client by concern (error, config, waf, client)
- Extract HttpError/HttpResult to error.rs (52 lines)
- Extract HttpClientConfig to config.rs (89 lines)
- Extract WAF detection to waf.rs (127 lines)
- Reduce client.rs to core implementation (330 lines)
- mod.rs re-exports public API
- All 448 tests passing, cargo check PASS
